+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 - New output `slack_reaction`. (@rockwotj) 
+- New processor `side_effect`. (@rockwotj) 
 
 ## 4.57.0 - 2025-06-10